Forth lesson
