--- lesson/user/exception/DataProcessingException.java
@@ -1,9 +1,0 @@
-package lesson.user.exception;
-
-import java.sql.SQLException;
-
-public class DataProcessingException extends RuntimeException {
-  public DataProcessingException(String message, SQLException e) {
-    super(message);
-  }
-}
