--- lesson/user/exception/EmailException.java
@@ -1,7 +1,0 @@
-package lesson.user.exception;
-
-public class EmailException extends RuntimeException {
-    public EmailException(String message) {
-        super(message);
-    }
-}
