--- lesson3/Grade.java
@@ -1,4 +1,0 @@
-package lesson3;
-
-public record Grade(String name, double score) {
-}
