--- lesson3/StudentGradeRecord.java
@@ -1,14 +1,0 @@
-package lesson3;
-
-public record StudentGradeRecord(String studentName, String schoolName, String subjectName, double score) {
-
-    @Override
-    public String toString() {
-        return "StudentGradeRecord{" +
-                "studentName='" + studentName + '\'' +
-                ", schoolName='" + schoolName + '\'' +
-                ", subjectName='" + subjectName + '\'' +
-                ", score=" + score +
-                '}';
-    }
-}
