+++ main/java/com/bchupika/lesson3/Address.java
@@ -1,0 +1,3 @@
+package com.bchupika.lesson3;
+
+public record Address(String street, String city) {}
