+++ main/java/com/bchupika/lesson3/Grade.java
@@ -1,0 +1,4 @@
+package com.bchupika.lesson3;
+
+public record Grade(String name, double score) {
+}
