+++ main/java/com/bchupika/user/exception/EmailException.java
@@ -1,0 +1,7 @@
+package com.bchupika.user.exception;
+
+public class EmailException extends RuntimeException {
+    public EmailException(String message) {
+        super(message);
+    }
+}
