+++ main/java/com/bchupika/user/exception/PasswordNotEqual.java
@@ -1,0 +1,7 @@
+package com.bchupika.user.exception;
+
+public class PasswordNotEqual extends RuntimeException {
+  public PasswordNotEqual(String message) {
+    super(message);
+  }
+}
