+++ main/java/com/bchupika/user/exception/PhoneNumberException.java
@@ -1,0 +1,7 @@
+package com.bchupika.user.exception;
+
+public class PhoneNumberException extends RuntimeException {
+    public PhoneNumberException(String message) {
+        super(message);
+    }
+}
