+++ main/java/com/bchupika/user/service/UserService.java
@@ -1,0 +1,10 @@
+package com.bchupika.user.service;
+
+import com.bchupika.user.dto.UserRegistrationDto;
+import com.bchupika.user.dto.UserResponseDto;
+
+interface UserService {
+    UserResponseDto registerUser(UserRegistrationDto userRegistrationDto);
+
+    UserResponseDto getUserById(Long userId);
+}
